+++ config.py
@@ -10,6 +10,8 @@ from pathlib import Path
 
 
 APP_NAME = "commsman"
+VERSION = "0.3.2"
+AUTHOR = "Carlosman1996"
 
 
 if hasattr(sys, '_MEIPASS'):